\name{abline.rv}
\title{Random Vectors}
\alias{abline.rv}
\usage{
 abline.rv(x, sims=1:20, \dots)
}
\description{
  \code{abline.rv} plots \code{n} lines corresponding to \code{n}
  first simulations of rv object \code{x}
}
\arguments{
  \item{x}{an object}
  \item{sims}{number of lines to draw}
  \item{\dots}{further arguments passed to or from other methods}
}
\details{
  ...
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

