% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_add.R
\name{add}
\alias{add}
\title{Add a Dependency}
\usage{
add(pkgs, dev = FALSE)
}
\arguments{
\item{pkgs}{Character vector. Names of packages to add. Supports CRAN names or "user/repo" for GitHub.}

\item{dev}{Logical. If \code{TRUE}, adds packages to "Suggests". If \code{FALSE} (default), adds to "Imports".}
}
\description{
Adds a package to the \code{DESCRIPTION} file, installs it into the local library using \code{pak},
and updates the \code{renv.lock} file.
}
