\name{as.rv.stanfit}
\alias{as.rv.stanfit}
\title{
  Convert simulations generated by Stan to a list of rv objects
}
\description{
  Convert simulations generated by Stan to a list of rv objects. 
}
\usage{
  \method{as.rv}{stanfit}(x, list.=TRUE, ...)
}
\arguments{
  \item{x}{A 'stanfit' object}
  \item{list.}{logical; return a list of \code{rv} objects instead of a
    single \code{rv} object (vector)?}
  \item{\dots}{(ignored)}
}
%\details{
%
%}
\value{
 A list of rv objects, with the names set for each rv object.
}
\references{
   Stan: \url{http://mc-stan.org/}
}
\author{
 J Kerman
}
%\note{
%%%  ~~further notes~~
%}
%%\seealso{
%%
%%}
\keyword{manip}
