% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvsummary.R
\name{as.rvsummary}
\alias{as.rvsummary}
\title{Coerce an object to an rvsummary}
\usage{
as.rvsummary(x, ...)
}
\arguments{
\item{x}{an rv object}

\item{\dots}{arguments passed to base::summary}
}
\description{
Coerce an object to an rvsummary
}
