% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_missing_deps}
\alias{check_missing_deps}
\title{Check for Missing Dependencies}
\usage{
check_missing_deps()
}
\description{
Checks if \code{renv} are installed. If not, stops with an informative message.
}
\keyword{internal}
