% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvmatch.R
\name{\%*in*\%}
\alias{\%*in*\%}
\title{Test if in set}
\usage{
x \%*in*\% y
}
\arguments{
\item{x}{random vector, regular atomic vector, or \code{NULL}: the values to
be matched.}

\item{y}{random vector, regular atomic vector, or \code{NULL}: the
values to be matched against.}
}
\description{
Test if in set
}
