% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matmult_rv.R
\name{\%**\%}
\alias{\%**\%}
\title{Constant matrix times a rv vector}
\usage{
x \%**\% y
}
\arguments{
\item{x}{numeric or complex matrices or vectors.}

\item{y}{numeric or complex matrices or vectors.}
}
\description{
Constant matrix times a rv vector
}
