% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_init.R
\name{init}
\alias{init}
\title{Initialize an rv Project}
\usage{
init(
  path = ".",
  repos = if (Sys.info()[["sysname"]] == "Linux") {
     c(`CRAN-LINUX` =
    "https://packagemanager.posit.co/cran/__linux__/manylinux_2_28/latest", `CRAN-OTHERS`
    = "https://packagemanager.posit.co/cran/latest")
 } else {
     c(`CRAN-OTHERS` =
    "https://packagemanager.posit.co/cran/latest", `CRAN-LINUX` =
    "https://packagemanager.posit.co/cran/__linux__/manylinux_2_28/latest")
 }
)
}
\arguments{
\item{path}{Character string. Path to the new project directory. Defaults to current working directory.}

\item{repos}{Character vector. Repositories to use. Defaults to \code{getOption("repos")}.}
}
\description{
Sets up a directory as an \code{rv} project by creating a \code{DESCRIPTION} file (if missing),
initializing a bare \code{renv} environment, setting the snapshot type to "explicit",
and configuring \code{.Rprofile} and \code{.Renviron}.
}
