\name{lines.rv}
\title{Plot some realized series of random vectors}
\alias{lines.rv}
\usage{
  lines.rv(x, simrows=1:3, lcolors=c("blue", "red", "green", "orange", "yellow"), \dots)
}
\description{
  \code{lines.rv}
}
\arguments{
  \item{x}{an object}
\item{simrows}{which rows of the simulation matrix to plot}
\item{lcolors}{colors to use when plotting a series}
\item{\dots}{further arguments passed to \code{lines}}
}
\details{
 ...
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

%%% 