% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_intent_repos}
\alias{load_intent_repos}
\title{Load Repositories from DESCRIPTION}
\usage{
load_intent_repos()
}
\description{
Reads the repositories defined in \verb{Config/intent/repos/} and sets them
in the global \code{options(repos)}.
}
\keyword{internal}
