\name{matsimapply}
\title{Apply a two-argument function to simulations}
\alias{matsimapply}
\usage{
  matsimapply(X,Y,FUN, \dots)
}
\description{
  (Used internally)
}
\arguments{
  \item{X}{}
  \item{Y}{}
  \item{FUN}{}
  \item{\dots}{further arguments passed to or from other methods}
}
\details{
  \code{matsimapply} 
}
\value{
  A random variable object.
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

%%% fill in!
