\name{mcmcarray}
\title{Coerce an rv object into a three-way array}
\alias{mcmcarray}
\usage{
  mcmcarray(x)
}
\description{
  \code{mcmcarray}
}
\arguments{
  \item{x}{an rv object that was coerced from a three-way array}
}
\details{
  \code{mcmcarray} 
}
\value{
  A three-way array. 
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

%%% OBSOLETE! (or should it be?)
