% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{read_intent_config}
\alias{read_intent_config}
\title{Read and parse custom Config/intent fields from DESCRIPTION}
\usage{
read_intent_config(
  path_to_description = file.path(getwd(), "DESCRIPTION"),
  mandatory = NULL,
  permissive = NULL
)
}
\arguments{
\item{path_to_description}{Path to the DESCRIPTION file.}

\item{mandatory}{Character vector of mandatory top-level settings (e.g., "repos").}

\item{permissive}{Character vector of permissive top-level settings. If NULL, all are allowed.
Permissive fields are optional; they are only used to filter out fields not in
\code{union(mandatory, permissive)}.}
}
\value{
A nested list of settings.
}
\description{
Read and parse custom Config/intent fields from DESCRIPTION
}
\keyword{internal}
