% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_remove.R
\name{remove}
\alias{remove}
\title{Remove a Dependency}
\usage{
remove(pkgs)
}
\arguments{
\item{pkgs}{Character vector. Names of packages to remove.}
}
\description{
Removes a package from the \code{DESCRIPTION} file, uninstalls it from the library,
and updates the \code{renv.lock} file.
}
