\name{rvhist}
\title{Histogram of Distributions of Components of a Random Vector}
\alias{rvhist}
\usage{
  rvhist(x, main=NULL, \dots)
}
\description{
  \code{rvhist}
}
\arguments{
  \item{x}{an rv object}
  \item{main}{main title}
  \item{\dots}{further arguments passed to the hist function}
}
\details{
  \code{rvhist} 
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

