% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvinvchisq.R
\name{rvinvchisq}
\alias{rvinvchisq}
\title{Generate Random Variables from a Inverse-Chi-Square Sampling Model}
\usage{
rvinvchisq(n = 1, df, scale = 1)
}
\arguments{
\item{n}{integer: number of variables to generate}

\item{df}{degrees of freedom (may be random)}

\item{scale}{scale parameter (may be random)}
}
\value{
A random vector (rv object).
}
\description{
\code{rvinvchisq}
}
\details{
\code{rvinvchisq}
}
\examples{

  rvinvchisq(df=3, scale=2)

}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects. Statistics and
Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
Jouni Kerman \email{jouni@kerman.com}
}
\keyword{classes}
