% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvmatch.R
\name{rvmatch}
\alias{rvmatch}
\title{Generate a Random Vector from a Bernoulli Sampling Model}
\usage{
rvmatch(x, table, nomatch = NA_integer_, incomparables = NULL)
}
\arguments{
\item{x}{random vector, regular atomic vector, or \code{NULL}: the values to
be matched.}

\item{table, y}{random vector, regular atomic vector, or \code{NULL}: the
values to be matched against.}

\item{nomatch}{the value to be returned in the case when no match is found.
Note that the value is coerced to \code{integer}.}

\item{incomparables}{a vector of values that cannot be matched.  Any value
in x matching a value in this vector is assigned the nomatch value.  For
historical reasons, FALSE is equivalent to \code{NULL}}
}
\value{
A random vector (an rv object) of the same length as \code{x}.

\code{rvmatch} returns an integer-valued vector.

\code{\%*in*\%} returns a logical-valued vector.

Both functions are compatible with regular atomic vectors.
}
\description{
\code{rvmatch} returns a (random) vector of the positions of (first) matches
of its first argument in its second.
}
\details{
\code{\%*in*\%} is a binary operator (analogous in its operation to
\code{\%in\%}) which returns a logical (random) vector indicating if there is
a match or not for its left operand.

...
}
\examples{

  x <- rvempirical(5, 1:10)
  z <- rvmatch(x, table=1:3, nomatch=0L)

}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects. Statistics and
Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
Jouni Kerman \email{jouni@kerman.com}
}
\keyword{classes}
