\name{rvmatrix}
\title{Matrices and Arrays of Random Vectors}
\alias{rvmatrix}
\alias{rvarray}
\usage{
 rvmatrix(\dots)
 rvarray(\dots)
}
\description{
  \code{rvarray}...
}
\arguments{
  \item{\dots}{arguments passed to \code{matrix} or to \code{array}}
}
\details{
  \code{rvarray} ...
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

