\name{rvmin}
\title{Minimum Simulated Value of Random Variable}
\alias{rvmin}
\usage{
rvmin(x, \dots)
}
\description{
  \code{rvmin}
}
\arguments{
  \item{x}{an object}
\item{\dots}{further arguments passed to or from other methods}
}
\details{
  \code{rvmin} 
}
\value{
  A \emph{numeric} vector of the same dimension as \code{x}.
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\seealso{
  \code{\link{rvmax}}, \code{\link{rvmedian}}, \code{\link{rvmean}}.
}
\examples{
  #
}

\keyword{classes}

