% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcattr.R
\name{rvnchains}
\alias{rvnchains}
\title{Number of Markov Chains Used to Generate Simulations of a Random Vector}
\usage{
rvnchains(x)
}
\arguments{
\item{x}{an rv object (supposed to be generated by a MCMC process)}
}
\description{
Retrieves the number of mcmc chains in each components of the argument.
}
\details{
Assumes that the rv object was generated by a MCMC process.  Umacs and
R2WinBUGS are compatible.
}
\examples{

  #

}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects. Statistics and
Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\seealso{
\code{\link{as.rv.bugs}}
}
\author{
Jouni Kerman \email{jouni@kerman.com}
}
\keyword{classes}
