\name{rvpar}
\title{Set or Query Parameters of the 'rv' Package}
\alias{rvpar}
\usage{
  rvpar(\dots)
}
\description{
  Retrieves the number of effective draws in each component of the argument.
}
\arguments{
  \item{\dots}{arguments in tag = value form, or a list of tagged values. The tags must come from the parameters described below.}
}
\details{
  %\item{oorm}{out-of range marker used in \code{mlplot} (experimental - not yet implemented)}
  \item{rvcol}{color of a random point}
  \item{rvlex}{...}
  \item{rvlwd}{line weight of a random interval}
  \item{print.digits}{number of digits to show in the summaries}
}
\value{
    
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
%\examples{
%}

\keyword{classes}

%%%
