% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvmin_rvmax_rvrange.R
\name{rvrange}
\alias{rvrange}
\title{Get the value range of an rv object}
\usage{
rvrange(x)
}
\arguments{
\item{x}{an \code{rv} or \code{rvsummary} object}
}
\description{
Get the value range of an rv object
}
