% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_rv.R
\name{sd.rv}
\alias{sd.rv}
\title{Get the standard deviation of an rv object}
\usage{
sd.rv(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an \R object which is coercible to one
    by \code{as.double(x)}.}

\item{na.rm}{logical.  Should missing values be removed?}
}
\description{
Get the standard deviation of an rv object
}
