% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_defaults.r
\name{show_RV_builtins}
\alias{show_RV_builtins}
\title{Show RV builtin rules}
\usage{
show_RV_builtins(rules = c("all", "validation", "cross"))
}
\arguments{
\item{rules}{Which builtin rules to show. \code{"validation"} for the
standard schema/data validation rules, \code{"cross"} for the cross rules that
check for consistency between rules, or \code{"all"} to show both.}
}
\value{
List of data.fames if \code{rules = "all"}, otherwise a single data.frame,
with information on the builtin rules for the rule type specified.
The data.frame(s) have an attached class \code{RV_rule_info}.
}
\description{
Helper to show the builtin rules in a Registry, and their behaviour.
}
\note{
The \code{RV_rule_info} class has a custom print method that formats the
data.frame(s) in a more readable way.
}
\examples{
show_RV_builtins()
}
\seealso{
\link{Registry} and \link{add_rule}.
}
