\name{simmapply}
\title{Random Vectors}
\alias{simmapply}
\usage{
  simmapply(FUN, X, \dots)
}
\description{
  apply a function to one or more random variables as arguments
}
\arguments{
  \item{FUN}{}
\item{X}{}
\item{\dots}{further arguments passed to or from other methods}
}
\details{
  \code{simmapply} 
}
\references{ 
Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations 
Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

