% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_sync.R
\name{sync}
\alias{sync}
\title{Sync the Environment}
\usage{
sync()
}
\description{
Ensures the \code{DESCRIPTION} file matches the \code{renv.lock} file exactly.
This is equivalent to \code{renv::restore(clean = TRUE)}.
}
