% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_rv.R
\name{var.rv}
\alias{var.rv}
\title{Calculate the variance of an rv object}
\usage{
var.rv(x, ...)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{\dots}{arguments passed to stats::var}
}
\description{
Calculate the variance of an rv object
}
