% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.rv.stanfit.R
\name{as.rv.stanfit}
\alias{as.rv.stanfit}
\title{Convert simulations generated by Stan to a list of rv objects}
\usage{
\method{as.rv}{stanfit}(x, list. = TRUE, ...)
}
\arguments{
\item{x}{A 'stanfit' object}

\item{list.}{logical; return a list of \code{rv} objects instead of a single
\code{rv} object (vector)?}

\item{\dots}{(ignored)}
}
\value{
A list of rv objects, with the names set for each rv object.
}
\description{
Convert simulations generated by Stan to a list of rv objects.
}
\references{
Stan: \url{https://mc-stan.org/}
}
\author{
J Kerman
}
\keyword{manip}
