% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_rbind_rv.R
\name{cbind.rv}
\alias{cbind.rv}
\alias{rbind.rv}
\title{Combine random vectors by columns or rows}
\usage{
\method{cbind}{rv}(..., deparse.level = 1)
}
\arguments{
\item{\dots}{vectors or matrices, can be rv objects}

\item{deparse.level}{(passed on to cbind)}
}
\description{
Combines random vectors by columns (\code{cbind.rv}) or rows
(\code{rbind.rv}).
}
\details{
See \link{cbind} and \link{rbind} for details.
}
\examples{
  x <- rvnorm(10)
  y <- rvnorm(10)
  cbind(x, y)
  rbind(x, y) 
}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects. Statistics and
Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
Jouni Kerman \email{jouni@kerman.com}
}
\keyword{classes}
