% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_rv.R
\name{cc}
\alias{cc}
\title{Combine values in an rv object}
\usage{
cc(..., recursive = FALSE)
}
\arguments{
\item{...}{objects to be concatenated.}

\item{recursive}{logical.  If \code{recursive = TRUE}, the function
    recursively descends through lists (and pairlists) combining all
    their elements into a vector.}
}
\description{
Combine values in an rv object
}
