% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_rv.R
\name{cov.rv}
\alias{cov.rv}
\title{Calculate the covariance of an rv object}
\usage{
cov.rv(x, y = NULL, ...)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{y}{\code{NULL} (default) or a vector, matrix or data frame with
    compatible dimensions to \code{x}.   The default is equivalent to
    \code{y = x} (but more efficient).}

\item{\dots}{arguments passed to stats::cov}
}
\description{
Calculate the covariance of an rv object
}
