% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachrv.R
\name{detachrv}
\alias{detachrv}
\title{Detach the rv package}
\usage{
detachrv()
}
\description{
\code{detachrv} detaches the rv package and restores the original functions
in \code{base}, \code{graphics} and \code{stats} packages.
}
\details{
Currently equivalent to \code{detach("package:rv")}.
}
\examples{

  \dontrun{
   library(rv)
   detachrv()
  }

}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects. Statistics and
Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
Jouni Kerman \email{jouni@kerman.com}
}
\keyword{classes}
