% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize an intent Project}
\usage{
init(
  path = ".",
  repos = if (Sys.info()[["sysname"]] == "Linux") {
     c(CRAN =
    "https://packagemanager.posit.co/cran/__linux__/manylinux_2_28/latest")
 } else {

     c(CRAN = "https://packagemanager.posit.co/cran/latest")
 }
)
}
\arguments{
\item{path}{Character string. Path to the new project directory. Defaults to current working directory.}

\item{repos}{Character vector. Repositories to use. Defaults to \code{getOption("repos")}.}
}
\description{
Sets up a directory as an \code{intent} project by creating a \code{DESCRIPTION} file (if missing),
initializing a bare \code{renv} environment, setting the snapshot type to "explicit",
and configuring \code{.Rprofile} and \code{.Renviron}.
}
