% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_Registry.r
\name{Registry}
\alias{Registry}
\alias{is.Registry}
\title{Create a Registry Object}
\usage{
Registry()

is.Registry(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
An S7 \code{Registry} object with the following properties:
\describe{
\item{\code{rule_names}}{All rule names, in order of evaluation.
Derived from \code{control}/\code{transform}/\code{validate} rules with
\code{apply_last} added to the end.}
\item{'control_rules}{Rules that influence the control flow.
These will be applied in the first pass when validating data.}
\item{'transform_rules'}{Rules that transform data.
These will be applied in the second pass when validating data.}
\item{'validate_rules'}{Rules that validate data.
These will be applied in the penultimate pass when validating
(the last pass is the special \code{apply_last} pass - see details).}
\item{\code{str_to_fn_rules}}{Schema rules that are allowed to have string
or function values, with string values being converted to functions
automatically during schema validation.}
\item{\code{str_to_fn_converter}}{A function that converts string values to
functions for the \code{str_to_fn_rules}.}
\item{\code{type_names}}{A character vector of allowed type names, derived
from the keys of the \code{type_map} environment.}
\item{\code{type_map}}{An environment mapping type names to type definition
functions.}
\item{\code{coerce_names}}{A character vector of allowed coercion rule names,
derived from the keys of the \code{coerce_map} environment.}
\item{\code{coerce_map}}{An environment mapping type names to coercion
functions.}
\item{\code{rule_names}}{A character vector of all (non-cross) rule names.}
\item{\code{schema_rules}}{An environment containing the schema rule
functions.}
\item{\code{cross_rule_names}}{A character vector of allowed schema cross
rule names, derived from the keys of the \code{cross_rules}
environment.}
\item{\code{cross_rules}}{An environment containing the schema cross
rule functions (rules that check relationships between multiple schema
rule values).}
\item{\code{validator_rules}}{An environment containing the validator rule
functions.}
}
}
\description{
An S7 class that defines rules for validating data against a \link{Schema}.
}
\details{
The \code{Registry} class serves as a central repository for all the rules used in
schema and data validation for \code{RV} based workflows. It includes builtin
rules for common validation tasks and allows for the addition of custom
rules.

Registry objects are automatically created in \code{Schema} objects
(which are passed to \code{Validator} objects), and rules can be added to the
registry directly within those objects by using the rule-adding generic
functions. It is therefore not necessary to create a \code{Registry} object
separately if one does not wish.

However, \code{Schema} and \code{Validator} objects are automatically re-validated
when rules are added, so for the addition of many new rules, it is
beneficial to create a \code{Registry} object, add all the rules to it, and
then pass it to the other classes.

For full details see the helper function \code{\link[=show_RV_builtins]{show_RV_builtins()}} or the
\href{../doc/RV-rules.html}{rules vignette}.
}
\examples{
r <- Registry()
s <- Schema(list(type = "integer"), registry = r)

is.Registry(r)
}
\seealso{
\link{Schema} and \link{Validator} constructors. \link{add_rule} for adding
rules to a registry.
}
