\name{Extract.rv}
\title{Extract or Replace Parts of a Random Vector}
\alias{[.rv}
\alias{[<-}
\alias{[<<-}
\alias{[<-.rv}
\alias{[<<-.rv}
\alias{$.rv}
\alias{$<-.rv}
\synopsis{
x[i]
x[i, j, \dots , drop = TRUE]
x$name
}
\description{
  %...
}
\details{
  Extracting rv objects works the same way as extracting
  components of a numerical vector or array.
  The return value is always an object of class 'rv'.
  Type ?Extract for details.

}
\value{
 A random variable (an rv object).
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}

\examples{
  #
}
\keyword{classes}

