\name{rv-package}
\alias{rv-package}
\docType{package}
\title{Simulation-based Random Variable Objects}
\description{
  `\code{rv}' implements a simulation-based random variable object class.

  Please refer to the vignette: \code{vignette("rv")} for details.
}
\details{
  \tabular{ll}{
   Package:   \tab rv \cr
   Version:   \tab 1.0 \cr
   Date:      \tab 2008/08/08 \cr
   Namespace: \tab rv \cr
   Enhances:  \tab R2WinBUGS \cr
   Depends:   \tab R(>= 2.5.1), methods, utils \cr
   License:   \tab GPL-2 \cr
  }
}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects.
Statistics and Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
 Jouni Kerman
 \email{jouni@kerman.com}
}
\keyword{classes}
