\name{rv-package}
\alias{rv-package}
\docType{package}
\title{Simulation-based Random Variable Objects}
\description{
  `\code{rv}' implements a simulation-based random variable object class.

  Please refer to the vignette: \code{vignette("rv")} for details.
}
\details{
  \tabular{ll}{
   Package:   \tab rv \cr
   Version: \tab 2.0.0 \cr
   Date: \tab 2012-01-27 \cr
   Namespace: \tab rv \cr
   Depends:   \tab R(>= 2.10.0), methods, utils, grDevices, graphics \cr
   License:   \tab GPL-2 \cr
  }
}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects.
Statistics and Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
 Jouni Kerman
 \email{jouni@kerman.com}
  Package built on Fri Jan 27 01:10:32 CET 2012
}
\keyword{classes}
