% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv-package.R
\docType{package}
\name{rv-package}
\alias{rv}
\alias{rv-package}
\title{rv: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Xiaochang Liu \email{liuxiaochang1996@gmail.com}

}
\keyword{internal}
