% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv.R
\docType{package}
\name{rv}
\alias{rv}
\alias{rv-package}
\title{rv: administrative boundary vector data}
\description{
This package downloads administrive boundaries from naturalearth and tigris
}
\section{Functions}{

\code{\link{v}}
}

