\name{rvattach}
\title{Attach / detach elements of rv objects to search path}
\alias{rvattach}
\usage{
rvattach(what, name='rvattach', \dots)
}
\description{
  
  \code{rvattach} 
}
\arguments{
  \item{what}{}
  \item{name}{}
  \item{\dots}{further arguments passed to or from other methods}
}
\details{
  \code{rvattach} 
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

%%%%