\name{rvattr}
\title{Attributes of Random Variables}
\alias{rvattr}
\alias{rvattr<-}
\usage{
rvattr(x, attrib)
rvattr(x, attrib) <- value
}
\description{
  \code{rvattr}
}
\arguments{
  \item{x}{an object}
  \item{attrib}{name of the attribute}
  \item{value}{value to be set}
}
\details{
  \code{rvattr} 
}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects.
Statistics and Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
 Jouni Kerman
 \email{jouni@kerman.com}
}
\examples{
  #
}

\keyword{classes}

