\name{rvattr}
\title{Attributes of Random Variables}
\alias{rvattr}
\alias{rvattr<-}
\usage{
rvattr(x, attrib)
rvattr(x, attrib) <- value
}
\description{
  \code{rvattr}
}
\arguments{
  \item{x}{an object}
  \item{attrib}{name of the attribute}
  \item{value}{value to be set}
}
\details{
  \code{rvattr} 
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

