\name{rvbeta}
\title{Generate Random Vectors from a Beta Sampling Model}
\alias{rvbeta}
\usage{
  rvbeta(n=1, shape1, shape2)
}
\description{
  \code{rvbeta}
}
\arguments{
  \item{n}{integer, number of random variables to generate}
  \item{shape1}{positive number or rv, 1st shape parameter}
  \item{shape2}{positive number or rv, 2nd shape parameter}
}
\details{
  \code{rvbeta} 
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  
}

\keyword{classes}

