% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvcov.R
\name{rvcov}
\alias{rvcov}
\title{Covariance Between Components of Random Vectors}
\usage{
rvcov(x, y = NULL, ...)
}
\arguments{
\item{x}{a random vector}

\item{y}{(optional) a random vector}

\item{\dots}{further arguments passed to or from other methods}
}
\value{
A covariance matrix.
}
\description{
\code{rvcov}
}
\details{
\code{rvcov}
}
\examples{

  x <- rvnorm(mean=1:3)
  y <- rvnorm(mean=2:4)
  rvcov(x,y)
  rvcov(x,x)

}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects. Statistics and
Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
Jouni Kerman \email{jouni@kerman.com}
}
\keyword{classes}
