% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvhist.R
\name{rvhist}
\alias{rvhist}
\title{Histogram of Distributions of Components of a Random Vector}
\usage{
rvhist(x, ...)
}
\arguments{
\item{x}{an rv object}

\item{\dots}{further arguments passed to the function \code{hist}}
}
\description{
\code{rvhist} shows a grid of histograms of simulations of the components of
a random vector.
}
\details{
Outputs a histogram using the \code{hist} function with the option
\code{freq=FALSE}. This can be overridden by specifying the argument
\code{freq} or \code{prob}.  See the function \code{hist} for details.
}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects. Statistics and
Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
Jouni Kerman \email{jouni@kerman.com}
}
\keyword{classes}
