\name{rvhist}
\title{Histogram of Distributions of Components of a Random Vector}
\alias{rvhist}
\usage{
  rvhist(x, \dots)
}
\description{
  \code{rvhist} shows a grid of histograms of simulations of the components
  of a random vector.
}
\arguments{
  \item{x}{an rv object}
  \item{\dots}{further arguments passed to the function \code{hist}}
}
\details{
}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects.
Statistics and Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
 Jouni Kerman
 \email{jouni@kerman.com}
}
\examples{
  #
}

\keyword{classes}

