\name{rvinci}
\title{Is a value inside a confidence interval?}
\alias{rvinci}
\usage{
  rvinci(obj, x, interval=0.95, one.sided=FALSE, left=FALSE)
}
\description{
  \code{rvinci} returns \code{TRUE} iff a given value is inside
  a given interval.
}
\arguments{
  \item{obj}{}
\item{x}{an object}
\item{interval}{}
\item{one.sided}{}
\item{left}{}
}
\details{
  
}
\value{
  A logical vector.
}
\references{ 
Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations 
Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

