\name{rvinvchisq}
\title{Generate Random Variables from a Inverse-Chi-Square Sampling Model}
\alias{rvinvchisq}
\usage{
  rvinvchisq(n=1, df, scale=1)
}
\description{
  \code{rvinvchisq}
}
\arguments{
  \item{n}{integer: number of variables to generate}
  \item{df}{degrees of freedom (may be random)}
  \item{scale}{scale parameter (may be random)}
}
\details{
  \code{rvinvchisq} 
}
\value{
  A random vector (rv object).
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

