% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvmin_rvmax_rvrange.R
\name{rvmax}
\alias{rvmax}
\title{Get the max values of an rv object}
\usage{
rvmax(x)
}
\arguments{
\item{x}{an \code{rv} or \code{rvsummary} object}
}
\description{
Get the max values of an rv object
}
