% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvmin_rvmax_rvrange.R
\name{rvmin}
\alias{rvmin}
\title{Get the min values of an rv object}
\usage{
rvmin(x)
}
\arguments{
\item{x}{an \code{rv} or \code{rvsummary} object}
}
\description{
Get the min values of an rv object
}
