\name{rvnchains}
\title{Number of Markov Chains Used to Generate Simulations of a Random Vector}
\alias{rvnchains}
\usage{
rvnchains(x)
}
\description{
  Retrieves the number of mcmc chains in each components of the argument.
}
\arguments{
  \item{x}{an rv object (supposed to be generated by a MCMC process)}
}
\details{
  Assumes that the rv object was generated by a MCMC process.
  Umacs and R2WinBUGS are compatible.
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.

Kerman, Jouni. Umacs: A Universal Markov Chain Sampler. Technical report, Columbia University, New York.
}
\seealso{
  \code{\link{as.rv.bugs}}
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

