% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcattr.R
\name{rvneff}
\alias{rvneff}
\title{Number of Effective Draws in Each Component of a Random Variable}
\usage{
rvneff(x)
}
\arguments{
\item{x}{an rv object}
}
\value{
A numeric object of the same length as the argument \code{x}.
}
\description{
Retrieves the number of effective draws in each component of the argument.
}
\details{
The number of effective draws is supposed to be saved by the simulation
generating program (e.g. WinBUGS via R2WinBUGS).
}
\examples{

  #

}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects. Statistics and
Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
Jouni Kerman \email{jouni@kerman.com}
}
\keyword{classes}
