\name{rvneff}
\title{Number of Effective Draws in Each Component of a Random Variable}
\alias{rvneff}
\usage{
  rvneff(x)
}
\description{
  Retrieves the number of effective draws in each component of the argument
  (supposed to have been saved by the simulation generating program).
}
\arguments{
  \item{x}{an object}
}
\details{
  \code{rvneff} 
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

%%%
