\name{rvpermut}
\title{Generate the Permutation Distribution of a Data Vector or Array}
\alias{rvpermut}
\usage{
  rvpermut(data)
}
\description{
  Generates a random vector that has the
  permutation distribution of a given constant vector or array.
}
\arguments{
  \item{data}{a constant vector or array}
}
\details{
  \code{rvpermut} 
}
\value{
  A random vector or array of the same dimension as \code{data}.
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

