\name{rvsimapply}
\title{Apply a Function to Columns of the Matrix of Simulation of a Random Variable}
\alias{rvsimapply}
\usage{
rvsimapply(x, FUN, \dots)
}
\description{
  \code{rvsimapply}
}
\arguments{
  \item{x}{an object}
  \item{FUN}{}
  \item{\dots}{further arguments passed to the function \code{FUN}}
}
\details{
  \code{rvsimapply} 
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

