% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvunif.R
\name{rvunif}
\alias{rvunif}
\title{Generate Random Vectors from a Uniform Sampling Model}
\usage{
rvunif(n = 1, min = 0, max = 1)
}
\arguments{
\item{n}{integer: number of scalars to generate}

\item{min}{lower limit of the distribution, (may be random)}

\item{max}{upper limit of the distribution, (may be random)}
}
\description{
Generates random variables from a Uniform sampling model.
}
\examples{

  y <- rvunif(1, min=rvunif(1)-1, rvunif(1)+1) # What marginal distribution does y have now?

}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects. Statistics and
Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
Jouni Kerman \email{jouni@kerman.com}
}
\keyword{classes}
