\name{solve.rv}
\title{Random Vectors}
\alias{solve.rv}
\usage{
  \method{solve}{rv}(a, b, \dots)
}
\description{
  \code{solve.rv}
}
\arguments{
 \item{a}{}
 \item{b}{}
 \item{\dots}{further arguments passed to or from other methods}
}
\details{
  \code{solve.rv} 
}
\references{ 
Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations 
Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

