% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_rv.R
\name{solve.rv}
\alias{solve.rv}
\title{Random Vectors}
\usage{
\method{solve}{rv}(a, b, ...)
}
\arguments{
\item{a}{a square random vector containing the coefficients of the linear
system}

\item{b}{a square random vector giving the right-hand side(s) of the linear
system}

\item{\dots}{further arguments passed to \code{solve}}
}
\description{
\code{solve.rv}
}
\details{
\code{solve.rv} is the rv-object compatible version of the function
\code{solve}.

For details of the function, see \code{\link{solve}}.
}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects. Statistics and
Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\seealso{
\code{\link{solve}}
}
\author{
Jouni Kerman \email{jouni@kerman.com}
}
\keyword{classes}
