\name{sort.rv}
\title{Distribution of Order Statistics of a Random Vector}
\alias{sort.rv}
\usage{
  \method{sort}{rv}(x, \dots)
}
\description{
  \code{sort.rv} computes the distribution of the order statistics
  of an rv object.
}
\arguments{
  \item{x}{an object}
  \item{\dots}{further arguments passed to \code{sort.rv}}
}
\value{
  An rv object of the same length as \code{x}.
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
%\seealso{
%  Other adapted functions: \code{\link{cov}}.
%}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

