\name{splitbyname}
\title{Split a vector based on the names of the components}
\alias{splitbyname}
\usage{
splitbyname(x)
}
\description{
  
  \code{splitbyname} splits the given vector based on the names
  of the components and returns a list.
  The names are supposed to be of the format 'name[indexes]',
  for example 'alpha[1,1]', 'beta[1]', etc.
  A name without brackets implies '[1]'.
}
\arguments{
  \item{x}{a vector with the name attributes set}
}
\details{
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{
  #
}

\keyword{classes}

%%%%
