\name{Summary.rv}
\title{Summary}
\alias{Summary.rv}
\usage{
  \method{Summary}{rv}(\dots, na.rm=FALSE)
}
\description{
  \code{Summary.rv}
}
\arguments{
  \item{\dots}{rv object}
  \item{na.rm}{logical, remove NAs?}
}
\details{
    
}
\references{Kerman, Jouni and Gelman, Andrew. Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Technical report, Columbia University, New York.
}
\author{
 Jouni Kerman
 \email{kerman@stat.columbia.edu}
 \url{http://www.stat.columbia.edu/~kerman}
}
\examples{

}

\keyword{internal}

