% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{Summary.rv}
\alias{Summary.rv}
\title{Summary}
\usage{
\method{Summary}{rv}(..., na.rm = FALSE)
}
\arguments{
\item{\dots}{rv object}

\item{na.rm}{logical, remove NAs?}
}
\description{
Methods applied to rv objects for the generic functions \code{all},
\code{any}, \code{sum}, \code{min}, \code{max}.
}
\examples{

  x <- rvnorm(5)
  all(x>0)
  any(x>2)
  sum(x)
  min(x)
  max(x)

}
\references{
Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing
Posterior Simulations Using Random Variable Objects. Statistics and
Computing 17:3, 235-244.

See also \code{vignette("rv")}.
}
\author{
Jouni Kerman \email{jouni@kerman.com}
}
\keyword{internal}
