% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v}
\alias{v}
\title{Select administrative boundary vector given a place code}
\usage{
v(country = NULL, fips = NULL, hasc = NULL, cb = TRUE,
  resolution = "20m", ...)
}
\arguments{
\item{country}{A country name identified in naturalearth country names}

\item{fips}{A FIPS code}

\item{hasc}{A HASC}

\item{cb}{Logical.  Applies only to FIPS codes.  If true will return a
generalized resolution cartographic boundary vector, otherwise returns a
higher resolution vector.  Default TRUE.}

\item{resolution}{The resolution if cb is TRUE.  Applies only to FIPS.
Options are 500k, 5m, and 20m for 1:500k, 1:5 million, and 1:20 million.}

\item{...}{Additional arguments passed to
\code{\link[rnaturalearth]{ne_states}},
\code{\link[rnaturalearth]{ne_countries}}, or
\code{\link[tigris]{counties}}.  Use `returnclass' (country and hasc) or
`class' (fips) to set the class of the returned geometry.  Both arguments
accept a value of either `sf' or `sp'.}
}
\value{
A vector for the administrative boundary identified by the location
  code
}
\description{
Accepts a place code and returns an administrative boundary vector.  This
function will service place codes from rnaturalearth place names, Federal
Information Processing Standard (FIPS) codes, and hierarchical administrative
subdivision codes (HASC).
}
\details{
Vector data is downloaded from rnaturalearth (place names and HASC) or tigris
(FIPS codes)
}
